Split shower_simulation and add Makefile
