+++ external/jsoncpp/jsoncpp.cpp
